cast return value of snprintf to size_t to make the compiler happy

+++ vlock-grab.c
@@ -71,7 +71,7 @@ int main(int argc, char **argv) {
   }
 
   /* format the virtual terminal filename from the number */
-  if (snprintf(vtname, sizeof vtname, VTNAME, vtno) > sizeof vtname) {
+  if ((size_t)snprintf(vtname, sizeof vtname, VTNAME, vtno) > sizeof vtname) {
     fprintf(stderr, "virtual terminal number too large\n");
   }
